implement getone search update

+++ controllers/backoffice/categorie.controller.js
@@ -7,7 +7,7 @@ const storingValidation = [
     body('categorieName').notEmpty()
 ]
 
-
+// Get All 
 const index = async (req, res) => {
     try {
         const categories = await Categorie.find({ deletedAt: null});
@@ -19,7 +19,7 @@ const index = async (req, res) => {
 
 }
 
-
+// Store a new Categorie
 const store = async (req, res) => {
     try {
         const errors = validationResult(req);
@@ -27,10 +27,16 @@ const store = async (req, res) => {
         {
             return res.status(400).json({ errors: errors.array() });
         }
-        const existingDeletedCategorie = await Categorie.findOne({ categorieName: req.body.categorieName, deletedAt:{$ne: null}})
+        const existingDeletedCategorie = await Categorie.findOne({ categorieName: req.body.categorieName})
+
 
         if(existingDeletedCategorie){
-            await existingDeletedCategorie.deleteOne();
+            if(existingDeletedCategorie.deletedAt === null){
+                res.status(409).json({ error: 'Categorie already exists' });
+                return
+            }else{
+                await existingDeletedCategorie.deleteOne();
+            }
         }
 
         let newCategorie = new Categorie({
@@ -49,4 +55,110 @@ const store = async (req, res) => {
 }
 
 
-module.exports = { index, store, storingValidation }
+// Get One categorie by Name
+const getOne = async (req, res) => {
+    try {
+        const categorie = await Categorie.findOne({ $and: [{categorieName: req.params.categoriename}, {deletedAt:null}] })
+        if(!categorie){
+            res.status(404);
+            res.json({
+                message: "Categorie Not found",
+                status: 404
+            })
+
+            return
+        }
+
+        return res.json(categorie)
+    } catch (error) {
+        res.json(internalError("", error));
+    }
+}
+
+
+// Search for categories
+const search = async (req, res) => {
+    const query = req.params.search;
+
+    try {
+        const categorie = await Categorie.find({
+            $or:[
+                { categorieName: { $regex: query, $options:'i' }}
+            ],
+            deletedAt: null
+        });
+
+        if(categorie.length === 0){
+            res.status(404);
+            res.json({
+                status: 404,
+                message: "No categorie founded"
+            });
+            return
+        }
+
+        res.json(categorie);
+    } catch (error) {
+        res.json(internalError("", error));
+    }
+}
+
+
+// update
+const update = async (req, res) => {
+    const id = req.params.id
+    try {
+        const errors = validationResult(req);
+        if(!errors.isEmpty()){
+            return res.status(400).json({ errors: errors.array() });
+        }
+
+        const categorie = await Categorie.findById(id);
+
+        if(!categorie)
+        {
+            res.status(404);
+            res.json({
+                status: 404,
+                message:"Categorie not found"
+            });
+            return
+        }
+
+        if(categorie.categorieName !== req.body.categorieName)
+        {
+            const sameName = await Categorie.findOne({
+                $and:[{
+                    categorieName: req.body.categorieName
+                },{
+                    _id: { $ne: id}
+                }]
+            });
+
+            if(sameName){
+                res.json({
+                    status: 401,
+                    messgae: "This Categorie already Exist"
+                });
+
+                return
+            }
+
+        }
+        categorie.categorieName = req.body.categorieName;
+        await categorie.save();
+        res.json({
+            data: categorie,
+            status: 200
+        });
+
+    } catch (error) {
+        res.json(internalError("", error));
+    }
+}
+
+const destroy = (req, res) => {
+
+}
+
+module.exports = { index, store,getOne,search, update, storingValidation }
